Make ingest endpoints accept generic fns instead of impl-specific args
In particular, instead of passing the business end of an MPSC queue to
the ingest listeners we now pass a generic Fn. Thus the ingest module
doesn't need to know about what (if any) queues we use elsewhere.

Additionally, this neatly solves the issue with feedback - previously
the handler would signal success to the source of the incoming Status
packet as soon as it was successfully passed on into the MPSC queue
without awaiting confirmation that it was successfully persisted. The
new interface allows us to ensure that ingest listeners actually get
notified in case of critical errors in the processing part of the
application.